requirements.txt
paho-mqtt
streamlit
pymongo
pandas

